+++ changelog.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Fixed the fileReadCsv and fileReadCsvAsStream APIs to maintain the order while reading CSV records](https://github.com/ballerina-platform/ballerina-standard-library/issues/3780)
 - [Fixed the issue related to expected type in CSV data mapping](https://github.com/ballerina-platform/ballerina-standard-library/issues/3669)
 - [Make some of the Java classes proper utility classes](https://github.com/ballerina-platform/ballerina-standard-library/issues/4901)
+- [Remove Error logger in module-ballerina-io](https://github.com/ballerina-platform/ballerina-standard-library/issues/3083)
 
 ## [1.3.1] - 2022-11-29
 ### Changed